Initial commit
This control unit is to control by switch 4 Servo controlled turnouts. The switch is 5 momentary Normal push-button type used to control the turnout direction. Point numbering is determined by the HO Layout design documents. These numbers are based on the sequential number within the scope of the panel that immediately address the point.